Deduplicate rate limit and concurrency helpers
