+++ src/handlers/ci-failure.ts
@@ -13,29 +13,9 @@ import {
   formatCISection,
   buildCIAnalysisMarker,
 } from "../lib/ci-failure-formatter.ts";
+import { mapWithConcurrency } from "../lib/concurrency.ts";
 
 const BASE_CHECK_FETCH_CONCURRENCY = 2;
-
-async function mapWithConcurrency<T, R>(
-  items: T[],
-  concurrency: number,
-  fn: (item: T, index: number) => Promise<R>,
-): Promise<R[]> {
-  const results = new Array<R>(items.length);
-  let nextIndex = 0;
-
-  async function worker(): Promise<void> {
-    while (nextIndex < items.length) {
-      const index = nextIndex;
-      nextIndex += 1;
-      results[index] = await fn(items[index]!, index);
-    }
-  }
-
-  const workerCount = Math.min(Math.max(1, concurrency), items.length);
-  await Promise.all(Array.from({ length: workerCount }, () => worker()));
-  return results;
-}
 
 async function fetchCheckRunsForRef(
   octokit: Awaited<ReturnType<GitHubApp["getInstallationOctokit"]>>,
