+++ src/lib/depends-impact-analyzer.ts
@@ -11,6 +11,7 @@
  */
 
 import type { Octokit } from "@octokit/rest";
+import { mapWithConcurrency } from "./concurrency.ts";
 import { withTimeBudget } from "./usage-analyzer.ts";
 
 const MODULE_CONTENT_FETCH_CONCURRENCY = 4;
@@ -41,27 +42,6 @@ export type ImpactResult = {
   timeLimitReached: boolean;
   degradationNote: string | null;
 };
-
-async function mapWithConcurrency<T, R>(
-  items: T[],
-  concurrency: number,
-  fn: (item: T, index: number) => Promise<R>,
-): Promise<R[]> {
-  const results = new Array<R>(items.length);
-  let nextIndex = 0;
-
-  async function worker(): Promise<void> {
-    while (nextIndex < items.length) {
-      const index = nextIndex;
-      nextIndex += 1;
-      results[index] = await fn(items[index]!, index);
-    }
-  }
-
-  const workerCount = Math.min(Math.max(1, concurrency), items.length);
-  await Promise.all(Array.from({ length: workerCount }, () => worker()));
-  return results;
-}
 
 // ─── Git Grep Output Parser ─────────────────────────────────────────────────
