+++ src/routes/slack-relay-webhooks.ts
@@ -6,6 +6,11 @@ import {
   evaluateWebhookRelayPayload,
   type NormalizedWebhookRelayEvent,
 } from "../slack/webhook-relay.ts";
+import {
+  createSlidingWindowRateLimiter,
+  requestSourceKey,
+  type RateLimitWindowOptions,
+} from "../lib/sliding-window-rate-limiter.ts";
 
 interface SlackRelayWebhookRouteDeps {
   config: AppConfig;
@@ -24,68 +29,6 @@ type SlackRelayRateLimitOptions = {
   preBody?: RateLimitWindowOptions;
   verified?: RateLimitWindowOptions;
 };
-
-type RateLimiter = {
-  isLimited(key: string): boolean;
-};
-
-function createSlidingWindowRateLimiter(
-  options: RateLimitWindowOptions | undefined,
-  defaults: Required<RateLimitWindowOptions>,
-): RateLimiter {
-  const max = options?.max ?? defaults.max;
-  const windowMs = options?.windowMs ?? defaults.windowMs;
-  const maxKeys = options?.maxKeys ?? defaults.maxKeys;
-  const timestampsByKey = new Map<string, number[]>();
-
-  function pruneKeys(cutoff: number): void {
-    if (timestampsByKey.size <= maxKeys) return;
-    for (const [key, timestamps] of timestampsByKey) {
-      if (timestamps.length === 0 || timestamps[timestamps.length - 1]! <= cutoff) {
-        timestampsByKey.delete(key);
-      }
-      if (timestampsByKey.size <= maxKeys) return;
-    }
-
-    for (const key of timestampsByKey.keys()) {
-      timestampsByKey.delete(key);
-      if (timestampsByKey.size <= maxKeys) return;
-    }
-  }
-
-  return {
-    isLimited(key: string): boolean {
-      const now = Date.now();
-      const cutoff = now - windowMs;
-      let timestamps = timestampsByKey.get(key);
-      if (!timestamps) {
-        timestamps = [];
-        timestampsByKey.set(key, timestamps);
-      }
-
-      const validStart = timestamps.findIndex((timestamp) => timestamp > cutoff);
-      if (validStart > 0) {
-        timestamps.splice(0, validStart);
-      } else if (validStart === -1) {
-        timestamps.length = 0;
-      }
-
-      if (timestamps.length >= max) {
-        pruneKeys(cutoff);
-        return true;
-      }
-
-      timestamps.push(now);
-      pruneKeys(cutoff);
-      return false;
-    },
-  };
-}
-
-function requestSourceKey(header: (name: string) => string | undefined): string {
-  const forwardedFor = header("x-forwarded-for")?.split(",")[0]?.trim();
-  return header("cf-connecting-ip") ?? header("x-real-ip") ?? forwardedFor ?? "unknown";
-}
 
 function secretsMatch(expected: string, provided: string | undefined): boolean {
   if (!provided) {
