+++ src/routes/webhooks.ts
@@ -7,6 +7,11 @@ import type { EventRouter, WebhookEvent } from "../webhook/types.ts";
 import type { RequestTracker, ShutdownManager, WebhookQueueStore } from "../lifecycle/types.ts";
 import { verifyWebhookSignature } from "../webhook/verify.ts";
 import { createChildLogger } from "../lib/logger.ts";
+import {
+  createSlidingWindowRateLimiter,
+  requestSourceKey,
+  type RateLimitWindowOptions,
+} from "../lib/sliding-window-rate-limiter.ts";
 
 interface WebhookRouteDeps {
   config: AppConfig;
@@ -30,68 +35,6 @@ type WebhookRateLimitOptions = {
   preBody?: RateLimitWindowOptions;
   verified?: RateLimitWindowOptions;
 };
-
-type RateLimiter = {
-  isLimited(key: string): boolean;
-};
-
-function createSlidingWindowRateLimiter(
-  options: RateLimitWindowOptions | undefined,
-  defaults: Required<RateLimitWindowOptions>,
-): RateLimiter {
-  const max = options?.max ?? defaults.max;
-  const windowMs = options?.windowMs ?? defaults.windowMs;
-  const maxKeys = options?.maxKeys ?? defaults.maxKeys;
-  const timestampsByKey = new Map<string, number[]>();
-
-  function pruneKeys(cutoff: number): void {
-    if (timestampsByKey.size <= maxKeys) return;
-    for (const [key, timestamps] of timestampsByKey) {
-      if (timestamps.length === 0 || timestamps[timestamps.length - 1]! <= cutoff) {
-        timestampsByKey.delete(key);
-      }
-      if (timestampsByKey.size <= maxKeys) return;
-    }
-
-    for (const key of timestampsByKey.keys()) {
-      timestampsByKey.delete(key);
-      if (timestampsByKey.size <= maxKeys) return;
-    }
-  }
-
-  return {
-    isLimited(key: string): boolean {
-      const now = Date.now();
-      const cutoff = now - windowMs;
-      let timestamps = timestampsByKey.get(key);
-      if (!timestamps) {
-        timestamps = [];
-        timestampsByKey.set(key, timestamps);
-      }
-
-      const validStart = timestamps.findIndex((timestamp) => timestamp > cutoff);
-      if (validStart > 0) {
-        timestamps.splice(0, validStart);
-      } else if (validStart === -1) {
-        timestamps.length = 0;
-      }
-
-      if (timestamps.length >= max) {
-        pruneKeys(cutoff);
-        return true;
-      }
-
-      timestamps.push(now);
-      pruneKeys(cutoff);
-      return false;
-    },
-  };
-}
-
-function requestSourceKey(header: (name: string) => string | undefined): string {
-  const forwardedFor = header("x-forwarded-for")?.split(",")[0]?.trim();
-  return header("cf-connecting-ip") ?? header("x-real-ip") ?? forwardedFor ?? "unknown";
-}
 
 export function createWebhookRoutes(deps: WebhookRouteDeps): Hono {
   const { config, logger, dedup, eventRouter, requestTracker, webhookQueueStore, shutdownManager } = deps;
